requirements.txt
dj-database-url==0.5.0
Django==1.9.6
django-easy-pdf==0.1.1
gunicorn==19.9.0
html5lib==1.0.1
Pillow==5.2.0
psycopg2==2.7.5
PyPDF2==1.26.0
reportlab==3.5.8
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.3
